Fix broken CaseNumberInput behavior
Jira ticket: CAMS-383
