+++ user-interface/src/data-verification/TransferOrderAccordion.test.tsx
@@ -16,7 +16,7 @@ import { describe } from 'vitest';
 import { orderType, orderStatusType } from '@/lib/utils/labels';
 import { MockData } from '@common/cams/test-utilities/mock-data';
 import { OfficeDetails } from '@common/cams/courts';
-import { getOfficeList, validateCaseNumberInput } from './dataVerificationHelper';
+import { getOfficeList } from './dataVerificationHelper';
 import { selectItemInMockSelect } from '@/lib/components/CamsSelect.mock';
 
 function isValidOrderTransfer(transfer: {
@@ -737,9 +737,8 @@ describe('TransferOrderAccordion', () => {
 
     enterCaseNumberInAccordion(caseIdInput, '00-00000');
     enterCaseNumberInAccordion(caseIdInput, '');
-    let approveButton: HTMLElement;
+    const approveButton = screen.getByTestId(`button-accordion-approve-button-${order.id}`);
     await waitFor(async () => {
-      approveButton = screen.getByTestId(`button-accordion-approve-button-${order.id}`);
       expect(approveButton).toBeInTheDocument();
       expect(approveButton).toBeVisible();
       expect(approveButton).toBeDisabled();
@@ -1009,49 +1008,5 @@ describe('Test CaseSelection component', () => {
     expect(document.body).toHaveTextContent(
       'USTP Office: transfer fromRegion ABC - Division Name 1toRegion BCD - Division Name 2',
     );
-  });
-});
-
-describe('Test validateCaseNumberInput function', () => {
-  beforeEach(async () => {
-    vi.stubEnv('CAMS_PA11Y', 'true');
-  });
-
-  test('When supplied a value with a length greater than 7, it should truncate value to 7 digits', async () => {
-    const testValue = '1234567890';
-    const resultValue = '12-34567';
-
-    const expectedResult = {
-      caseNumber: resultValue,
-      joinedInput: resultValue,
-    };
-
-    const testEvent = {
-      target: {
-        value: testValue,
-      },
-    };
-
-    const returnedValue = validateCaseNumberInput(testEvent as React.ChangeEvent<HTMLInputElement>);
-    expect(returnedValue).toEqual(expectedResult);
-  });
-
-  test('When supplied a value with alphabetic characters only, it should return an object with undefined caseNumber and empty string for joinedInput', async () => {
-    const testValue = 'abcdefg';
-    const resultValue = '';
-
-    const expectedResult = {
-      caseNumber: undefined,
-      joinedInput: resultValue,
-    };
-
-    const testEvent = {
-      target: {
-        value: testValue,
-      },
-    };
-
-    const returnedValue = validateCaseNumberInput(testEvent as React.ChangeEvent<HTMLInputElement>);
-    expect(returnedValue).toEqual(expectedResult);
   });
 });
