+++ user-interface/src/lib/components/CaseNumberInput.test.ts
@@ -1,5 +1,50 @@
 import { describe } from 'vitest';
+import { validateCaseNumberInput } from './CaseNumberInput';
 
 describe('Case number input component', () => {
   test('test validateCaseNumberInput()', () => {});
+});
+
+describe('Test validateCaseNumberInput function', () => {
+  beforeEach(async () => {
+    vi.stubEnv('CAMS_PA11Y', 'true');
+  });
+
+  test('When supplied a value with a length greater than 7, it should truncate value to 7 digits', async () => {
+    const testValue = '1234567890';
+    const resultValue = '12-34567';
+
+    const expectedResult = {
+      caseNumber: resultValue,
+      joinedInput: resultValue,
+    };
+
+    const testEvent = {
+      target: {
+        value: testValue,
+      },
+    };
+
+    const returnedValue = validateCaseNumberInput(testEvent as React.ChangeEvent<HTMLInputElement>);
+    expect(returnedValue).toEqual(expectedResult);
+  });
+
+  test('When supplied a value with alphabetic characters only, it should return an object with undefined caseNumber and empty string for joinedInput', async () => {
+    const testValue = 'abcdefg';
+    const resultValue = '';
+
+    const expectedResult = {
+      caseNumber: undefined,
+      joinedInput: resultValue,
+    };
+
+    const testEvent = {
+      target: {
+        value: testValue,
+      },
+    };
+
+    const returnedValue = validateCaseNumberInput(testEvent as React.ChangeEvent<HTMLInputElement>);
+    expect(returnedValue).toEqual(expectedResult);
+  });
 });
